move settingHelper
